drawPage и getElementFromString вынесены в utils.js
