--- js/draw-page.js
@@ -1,11 +1,0 @@
-export default (pageElement) => {
-
-  const mainSection = document.querySelector(`.app .main`);
-  if (mainSection) {
-    if (mainSection.firstElementChild) {
-      mainSection.replaceChild(pageElement, mainSection.firstElementChild);
-    } else {
-      mainSection.appendChild(pageElement);
-    }
-  }
-};
