--- js/get-element-from-string.js
@@ -1,5 +1,0 @@
-export default (str) => {
-  const element = document.createElement(`section`);
-  element.innerHTML = str;
-  return element.firstElementChild ? element.firstElementChild : false;
-};
